rename filterContainer to filter
