+++ src/components/views/projects/filter/ProjectsFiltersMobile.tsx
@@ -1,9 +1,9 @@
 import ProjectsSearchTablet from '@/components/views/projects/ProjectsSearchTablet';
 import { Flex } from '@/components/styled-components/Flex';
 import ProjectsSubCategories from '@/components/views/projects/ProjectsSubCategories';
-import ProjectsFiltersSwiper from '@/components/views/projects/filterContainer/ProjectsFiltersSwiper';
+import ProjectsFiltersSwiper from '@/components/views/projects/filter/ProjectsFiltersSwiper';
 import { StyledLine } from '@/components/views/projects/common.styled';
-import ProjectsFiltersButton from '@/components/views/projects/filterContainer/ProjectsFiltersButton';
+import ProjectsFiltersButton from '@/components/views/projects/filter/ProjectsFiltersButton';
 
 const ProjectsFiltersMobile = () => {
 	return (
